Cleaned up and refactored archiving code
