+++ config/sync/core.extension.yml
@@ -86,7 +86,6 @@ module:
   path: 0
   path_alias: 0
   phpass: 0
-  pluginformalter: 0
   quick_node_clone: 0
   redirect: 0
   responsive_image: 0
