+++ web/modules/custom/hoeringsportal_citizen_proposal_archiving/hoeringsportal_citizen_proposal_archiving.info.yml
@@ -1,10 +1,9 @@
 name: Hoeringsportal citizen proposal archiving
 type: module
 description: 'Archives citizen proposals in GetOrganized.'
-core_version_requirement: ^9 || ^10
+core_version_requirement: ^10
 package: ITK
 dependencies:
-  - drupal:advancedqueue
-  - drupal:entity_events
-  - drupal:pluginformalter
+  - advancedqueue:advancedqueue
+  - entity_events:entity_events
   - hoeringsportal_citizen_proposal:hoeringsportal_citizen_proposal
