+++ web/modules/custom/hoeringsportal_citizen_proposal_archiving/hoeringsportal_citizen_proposal_archiving.services.yml
@@ -1,25 +1,17 @@
 services:
+  _defaults:
+    autowire: true
+
   logger.channel.hoeringsportal_citizen_proposal_archiving:
     parent: logger.channel_base
     arguments: [ 'hoeringsportal_citizen_proposal_archiving' ]
 
   Drupal\hoeringsportal_citizen_proposal_archiving\Archiver\GetOrganizedArchiver:
-    arguments:
-      - '@file_system'
-      - '@database'
-      - '@logger.channel.hoeringsportal_citizen_proposal_archiving'
 
   Drupal\hoeringsportal_citizen_proposal_archiving\Renderer\Renderer:
-    arguments:
-      - '@renderer'
-      - '@module_handler'
 
   Drupal\hoeringsportal_citizen_proposal_archiving\Helper\Helper:
-    arguments:
-      - '@entity_type.manager'
-      - '@Drupal\hoeringsportal_citizen_proposal\Helper\Helper'
-      - '@Drupal\hoeringsportal_citizen_proposal_archiving\Archiver\GetOrganizedArchiver'
-      - '@Drupal\hoeringsportal_citizen_proposal_archiving\Renderer\Renderer'
-      - '@logger.channel.hoeringsportal_citizen_proposal_archiving'
     tags:
       - {name: event_subscriber}
+
+  Drupal\hoeringsportal_citizen_proposal_archiving\Helper\FormHelper:
